rename selenium_functions and split functions in differents file, now scrape_data.py only scrape and download the data
